rewrite CourseInstanceSyncErrorsAndWarnings in Preact
